+++ lib/models/transactions.dart
@@ -3,7 +3,7 @@ import 'package:flutter/foundation.dart';
 class Transaction {
   String? id;
   String? title;
-  double? amount;
+  int? amount;
   DateTime? date;
 
   Transaction(
